requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
fpdf2==2.5.1
idna==3.3
mysqlclient==2.1.0
Pillow==9.0.1
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
